Set outline-level function, and change depth to start at 1
Set outline-level in *Simple Call Tree* buffer new function
simple-call-tree-outline-level. Alter outline-regexp so that it
captures the arrow in a match-string which is then used to determine
the depth with simple-call-tree-outline-level.
Change all references and assignments to depth so that level 1
is the top level.

Altered items:
simple-call-tree-mode, simple-call-tree-display-buffer,
simple-call-tree-list-callers-and-functions,
simple-call-tree-list-callees-recursively,
simple-call-tree-outline-level, simple-call-tree-invert-buffer,
simple-call-tree-change-maxdepth