New functions for navigating jump-ring
New functions simple-call-tree-jump-ring-prev and
simple-call-tree-jump-ring-next for moving back and forth
through the jump-ring. Introduce new global variable
simple-call-tree-jump-ring-index to make this work properly.
Add keys to major-mode for navigating jump-ring.
Add arg to simple-call-tree-jump-to-function controlling whether
or not function is added to the jump-ring.

Altered items:
simple-call-tree-mode, simple-call-tree-jump-ring-index,
simple-call-tree-jump-to-function, simple-call-tree-jump-prev,
simple-call-tree-jump-next